fixing init database script not to drop database and adding to docker compose

+++ init_db.py
@@ -1,7 +1,9 @@
 import sys
+from argparse import ArgumentParser
+
 from app import create_app, db
 
-def initialize_database():
+def initialize_database(drop_all):
     """Drops existing tables and creates new ones based on the models."""
     app = create_app()
     
@@ -10,8 +12,9 @@ def initialize_database():
             print("--- Database Initialization Started ---")
             
             # 1. Clear existing data (Optional: remove if you want to persist data)
-            print("Dropping all existing tables...")
-            db.drop_all()
+            if drop_all:
+                print("Dropping all existing tables...")
+                db.drop_all()
             
             # 2. Create tables based on models/ definitions
             print("Creating new tables...")
@@ -24,5 +27,11 @@ def initialize_database():
             print(f"❌ Error during initialization: {e}", file=sys.stderr)
             sys.exit(1)
 
+def main():
+    parser = ArgumentParser(description="Initialize the database.")
+    parser.add_argument("--drop-all", action="store_true", help="Drop all existing tables before creating new ones.", default=False)
+    args = parser.parse_args()
+    initialize_database(drop_all=args.drop_all)
+
 if __name__ == "__main__":
-    initialize_database()
+    main()
